Refactor + nuevos cambios en deep learning
