+++ .gitignore
@@ -1,0 +1,2 @@
+SBW-vectors-300-min5.txt
+SBW-vectors-300-min5.txt.bz2
